preliminary template restructuring
Now, instead of separately compiling different templates and manually
nesting and embedding them, I'm leveraging the partial template system.
This somewhat simplifies the rendering process.

I intend to change the function that handlebars::render takes in the
third parameter to be one that yields a BTreeMap instead of a Json, to
encourage and enable the composition of different sources of data.